requirements.txt
beautifulsoup4==4.12.2
discord.py==2.3.2
ffmpeg==1.4
gTTS==2.3.2
PyNaCl==1.5.0
requests==2.31.0
# yt-dlp==2023.7.6
